refactor: move components to root component folder
