--- src/app/(application)/_components/navbar.tsx
@@ -1,12 +1,0 @@
-import { SidebarTrigger } from "@/components/ui/sidebar";
-import React from "react";
-
-const Navbar = () => {
-  return (
-    <div className="flex w-full items-center justify-between">
-      <SidebarTrigger />
-    </div>
-  );
-};
-
-export default Navbar;
