--- src/app/(application)/admin/subjects/layout.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Layout = ({ children }: { children: React.ReactNode }) => {
-  return (
-      <div>{children}</div>
-  );
-};
-
-export default Layout;
