--- src/app/(application)/layout.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Layout = ({ children }: { children: React.ReactNode }) => {
-  return <div>{children}</div>;
-};
-
-export default Layout;
