+++ src/components/info-bar.tsx
